// src/types/dataExtraction.ts
export const SectionTypeEnum = {
  FlatForm: 0,
  MatrixGrid: 1,
} as const;

export type SectionTypeEnum = typeof SectionTypeEnum[keyof typeof SectionTypeEnum];